Added MIT license.  Updated docs, prebuilt hex.

+++ App/t1_params.c
@@ -1,3 +1,27 @@
+/*
+ * Copyright (c) 2023-2024 i3 Product Development
+ * 
+ * MIT License
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+
 /********************************************************************************************
  *    _ ____  ___             _         _     ___              _                        _
  *   (_)__ / | _ \_ _ ___  __| |_  _ __| |_  |   \ _____ _____| |___ _ __ _ __  ___ _ _| |_
@@ -475,7 +499,7 @@ static int write_param_to_nvm(const uint32_t pid, const cr_ParameterValue *param
 
 
 
-int crcb_ping_get_signal_strength(int32_t *rssi)
+int crcb_ping_get_signal_strength(int8_t *rssi)
 {
     sl_bt_connection_get_median_rssi(rsl_get_connection(), rssi);
     return 0;
